moved some files around, completed web scraping, need to do the dictionary thing i think
