+++ CP9/lesson26/my_set.py
